Restructured search scripts for more common usecase
