+++ src/scripts/global/search/constant.js
@@ -1,0 +1,5 @@
+const DOCS_IDBVALIDKEY = 'docs__cache'
+
+export {
+  DOCS_IDBVALIDKEY
+}
